pyproject.toml
[project]
name = "avito-report-builder"
version = "0.1.0"
description = "Internal tool to process Avito Excel exports and generate standardized reports"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
